+++ src/holmes-rs/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Widened CemaNeige `qnbv` calibration bounds from `[50, 800]` to `[0, 2000]`
+
 ## [0.4.0] - 2026-04-29
 
 ### Added